requirements.txt
alpaca-py>=0.21.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
jinja2>=3.1.0
httpx>=0.27.0
aiohttp>=3.9.0
